fix(7448): export leaflet-pixi-overlay from videx-map
