+++ .storybook/src/PixiOverlay.d.ts
@@ -1,6 +1,10 @@
 declare namespace L {
   class PixiOverlay {
     constructor(drawCallback: any, container: any, options: any);
+
+    onAdd(map: L.Map): void;
+
+    onRemove(map: L.Map): void;
   }
 
   namespace LeafletEvent {
