+++ tsconfig.json
@@ -8,7 +8,8 @@
     "target": "ES2022",
     "esModuleInterop": true,
     "types": ["vitest/globals", "node"],
-    "resolveJsonModule": true
+    "resolveJsonModule": true,
+    "allowJs": true
   },
   "include": [
     "./src",
